Fuse data
